+++ scripts/versionbump.js
@@ -1,4 +1,3 @@
-const replace = require('replace-in-file');
 const newVersion = process.argv[2];
 
 if (!newVersion || !newVersion.match(/\d+\.\d+.\d+(?:-\w+(?:\.\w+)?)?/)) {
@@ -11,25 +10,29 @@ function ShowModifiedFiles(changedFiles) {
   console.log('Modified files:', files.join(', '));
 }
 
-replace({
-  files: ['./src/js/version.ts'],
-  from: /\d+\.\d+.\d+(?:-\w+(?:\.\w+)?)?/g,
-  to: newVersion,
-})
-  .then(changedFiles => {
+(async () => {
+  //New module no longer exports CommonJS format
+  const { replaceInFile } = await import('replace-in-file');
+  try {
+    const changedFiles = await replaceInFile({
+      files: ['./src/js/version.ts'],
+      from: /\d+\.\d+.\d+(?:-\w+(?:\.\w+)?)?/g,
+      to: newVersion,
+    });
+
     ShowModifiedFiles(changedFiles);
-    return replace({
+    const changedPluginFile = await replaceInFile({
       files: ['plugin.xml'],
       // from: /id="@sentry\/cordova" version="\d+\.\d+.\d+"/g,
       // to: `id="@sentry/cordova" version="${pjson.version}"`,
       from: /id="sentry-cordova" version="\d+.\d+.\d+[^"]*"/g,
       to: `id="sentry-cordova" version="${newVersion}"`,
     });
-  })
-  .then(changedFiles => {
-    ShowModifiedFiles(changedFiles);
-  })
-  .catch(error => {
+
+    ShowModifiedFiles(changedPluginFile);
+  }
+  catch (error) {
     console.error('Error occurred:', error);
     process.exit(1);
-  });
+  };
+})();